+++ AdNetworkSupport/InMobi/InMobiBannerCustomEvent.h
@@ -2,15 +2,14 @@
 //  InMobiBannerCustomEvent.h
 //  MoPub
 //
-//  Copyright (c) 2012 MoPub, Inc. All rights reserved.
+//  Copyright (c) 2013 MoPub, Inc. All rights reserved.
 //
 
 #import "MPBannerCustomEvent.h"
 #import "IMBanner.h"
-#import "IMInMobiNetworkExtras.h"
 
 /*
- * Certified with version 4.0.2 of the InMobi SDK.
+ * Certified with version 4.1.0 of the InMobi SDK.
  */
 
 @interface InMobiBannerCustomEvent : MPBannerCustomEvent <IMBannerDelegate>