+++ AdNetworkSupport/InMobi/InMobiInterstitialCustomEvent.h
@@ -2,17 +2,16 @@
 //  InMobiInterstitialCustomEvent.h
 //  MoPub
 //
-//  Copyright (c) 2012 MoPub, Inc. All rights reserved.
+//  Copyright (c) 2013 MoPub, Inc. All rights reserved.
 //
 
 #import "MPInterstitialCustomEvent.h"
 
 #import "IMInterstitial.h"
 #import "IMInterstitialDelegate.h"
-#import "IMInMobiNetworkExtras.h"
 
 /*
- * Certified with version 4.0.2 of the InMobi SDK.
+ * Certified with version 4.1.0 of the InMobi SDK.
  */
 
 @interface InMobiInterstitialCustomEvent : MPInterstitialCustomEvent <IMInterstitialDelegate>